+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 12.3
 - Fix PHP8.4 deprecated passing E_USER_ERROR to trigger_error() in curl.php & Markdownify/Converter.php
 - Fix PHP8.4 Implicitly marking parameter as nullable is deprecated in Parsedown.php
 - Security: Cache killer: use file last modified time hash instead of RosarioSIS version in Warehouse.php
+- Include PHP ctype extension emulation by Symfony in Ctype.php, PHPCompatibility.php & ctype.php
 
 Changes in 12.2.3
 -----------------